+++ path.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"fmt"
+	"os"
 )
 
 var cmdPath = &Command{
@@ -24,6 +25,10 @@ For more about how GOPATH works, see 'go help gopath'.
 func runPath(cmd *Command, args []string) {
 	if len(args) != 0 {
 		cmd.UsageExit()
+	}
+	if VendorExperiment {
+		fmt.Fprintln(os.Stderr, "Error: GOVENDOREXPERIMENT is enabled and the vendor/ directory is not a valid Go workspace.")
+		os.Exit(1)
 	}
 	gopath := prepareGopath()
 	fmt.Println(gopath)